Refactor and move instrumented tests for another package
